requirements.txt
pyqt5
pyserial
numpy
plotter
matplotlib
tikzplotlib